# Assignment 12/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SentinelPay Fraud Detection API",
    "description": "Real-time fraud detection and prevention REST API. Implements the Transaction ingestion pipeline (FR-01\u2013FR-08), Fraud Case management (FR-09\u2013FR-10), and ML Model lifecycle (FR-13\u2013FR-14) from the SentinelPay System Requirements Document.\n\n**Author:** Teboho Mokoni  \n**Assignment:** 12 \u2013 Service Layer and REST API  \n**Stack:** Python 3.12 \u00b7 FastAPI \u00b7 In-memory repositories (A11)",
    "contact": {
      "name": "Teboho Mokoni",
      "url": "https://github.com/Teboho66/SentinelPay"
    },
    "license": {
      "name": "Academic project \u2014 CPUT Postgraduate Diploma"
    },
    "version": "1.0.0"
  },
  "paths": {
    "/api/transactions": {
      "post": {
        "tags": [
          "Transactions"
        ],
        "summary": "Submit a transaction for fraud evaluation",
        "description": "Accepts a transaction from the Payment Processor, validates the schema (FR-02), tokenises PII (BR-T2), and persists the transaction ready for ML scoring.",
        "operationId": "submit_transaction_api_transactions_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitTransactionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transaction accepted and persisted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            }
          },
          "409": {
            "description": "Duplicate transaction_id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Schema validation failed (FR-02)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Transactions"
        ],
        "summary": "Fetch all transactions",
        "description": "Returns all persisted transactions. Supports optional filtering by decision or risk_tier.",
        "operationId": "get_all_transactions_api_transactions_get",
        "parameters": [
          {
            "name": "decision",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filter by FraudDecision",
              "title": "Decision"
            },
            "description": "Filter by FraudDecision",
            "example": "HARD_BLOCK"
          },
          {
            "name": "risk_tier",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filter by RiskTier",
              "title": "Risk Tier"
            },
            "description": "Filter by RiskTier",
            "example": "CRITICAL"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransactionResponse"
                  },
                  "title": "Response Get All Transactions Api Transactions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/transactions/{transaction_id}": {
      "get": {
        "tags": [
          "Transactions"
        ],
        "summary": "Get a transaction by ID",
        "operationId": "get_transaction_api_transactions__transaction_id__get",
        "parameters": [
          {
            "name": "transaction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Transaction Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Transaction not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Transactions"
        ],
        "summary": "Delete a transaction",
        "operationId": "delete_transaction_api_transactions__transaction_id__delete",
        "parameters": [
          {
            "name": "transaction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Transaction Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Transaction not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/transactions/{transaction_id}/decision": {
      "post": {
        "tags": [
          "Transactions"
        ],
        "summary": "Apply fraud decision to a transaction",
        "description": "FR-07: Applies ML ensemble model scores, computes the composite fraud_score, and produces a FraudDecision (APPROVE / SOFT_DECLINE / HARD_BLOCK) using per-account-tier thresholds.",
        "operationId": "apply_decision_api_transactions__transaction_id__decision_post",
        "parameters": [
          {
            "name": "transaction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Transaction Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyDecisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Decision applied successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Transaction not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Decision already applied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid model scores",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApplyDecisionRequest": {
        "properties": {
          "fraud_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Fraud Score",
            "example": 0.87
          },
          "model_scores": {
            "items": {
              "$ref": "#/components/schemas/ModelScoreInput"
            },
            "type": "array",
            "minItems": 1,
            "title": "Model Scores"
          },
          "account_tier": {
            "type": "string",
            "title": "Account Tier",
            "description": "One of: STANDARD, PREMIUM, BUSINESS",
            "default": "STANDARD",
            "example": "STANDARD"
          }
        },
        "type": "object",
        "required": [
          "fraud_score",
          "model_scores"
        ],
        "title": "ApplyDecisionRequest",
        "description": "POST /api/transactions/{id}/decision \u2014 apply ML ensemble scores and produce FraudDecision."
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "status_code": {
            "type": "integer",
            "title": "Status Code"
          }
        },
        "type": "object",
        "required": [
          "error",
          "detail",
          "status_code"
        ],
        "title": "ErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModelScoreInput": {
        "properties": {
          "model_name": {
            "type": "string",
            "title": "Model Name",
            "example": "XGBOOST"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version",
            "example": "3.1"
          },
          "raw_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Raw Score",
            "example": 0.87
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "example": 0.91
          }
        },
        "type": "object",
        "required": [
          "model_name",
          "model_version",
          "raw_score",
          "confidence"
        ],
        "title": "ModelScoreInput"
      },
...
      "SubmitTransactionRequest": {
        "properties": {
          "transaction_id": {
            "type": "string",
            "title": "Transaction Id",
            "description": "Unique transaction identifier",
            "example": "TXN-2026-001"
          },
          "account_id_token": {
            "type": "string",
            "title": "Account Id Token",
            "description": "Tokenised account identifier",
            "example": "acc_token_abc123"
          },
          "merchant_id": {
            "type": "string",
            "title": "Merchant Id",
            "example": "MER-001"
          },
          "merchant_category_code": {
            "type": "string",
            "title": "Merchant Category Code",
            "description": "ISO 18245 MCC",
            "example": "5411"
          },
          "amount": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Amount",
            "description": "Transaction amount (BR-T1: Decimal precision)",
            "example": 1500.0
          },
          "currency": {
            "type": "string",
            "maxLength": 3,
            "minLength": 3,
            "title": "Currency",
            "description": "ISO 4217 currency code",
            "example": "ZAR"
          },
          "channel": {
            "type": "string",
            "title": "Channel",
            "description": "One of: CNP_ONLINE, CNP_MOBILE, POS, ATM",
            "example": "CNP_ONLINE"
          },
          "device_fingerprint_token": {
            "type": "string",
            "title": "Device Fingerprint Token",
            "example": "dfp_token_xyz789"
          },
          "ip_address_hash": {
            "type": "string",
            "title": "Ip Address Hash",
            "example": "a3f8bc912d"
          },
          "latitude": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude",
            "example": -33.9249
          },
          "longitude": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude",
            "example": 18.4241
          },
          "is_international": {
            "type": "boolean",
            "title": "Is International",
            "default": false,
            "example": false
          }
        },
        "type": "object",
        "required": [
          "transaction_id",
          "account_id_token",
          "merchant_id",
          "merchant_category_code",
          "amount",
          "currency",
          "channel",
          "device_fingerprint_token",
          "ip_address_hash",
          "latitude",
          "longitude"
        ],
        "title": "SubmitTransactionRequest",
        "description": "POST /api/transactions \u2014 submit a new transaction for fraud evaluation."
      },
      "TransactionResponse": {
        "properties": {
          "transaction_id": {
            "type": "string",
            "title": "Transaction Id"
          },
          "account_id_token": {
            "type": "string",
            "title": "Account Id Token"
          },
          "merchant_id": {
            "type": "string",
            "title": "Merchant Id"
          },
          "merchant_category_code": {
            "type": "string",
            "title": "Merchant Category Code"
          },
          "amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "channel": {
            "type": "string",
            "title": "Channel"
          },
          "is_international": {
            "type": "boolean",
            "title": "Is International"
          },
          "fraud_score": {
            "type": "number",
            "title": "Fraud Score"
          },
          "risk_tier": {
            "type": "string",
            "title": "Risk Tier"
          },
          "decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision"
          },
          "model_version_composite": {
            "type": "string",
            "title": "Model Version Composite"
          },
          "processing_ms": {
            "type": "integer",
            "title": "Processing Ms"
          },
          "pii_tokenised": {
            "type": "boolean",
            "title": "Pii Tokenised"
          }
        },
        "type": "object",
        "required": [
          "transaction_id",
          "account_id_token",
          "merchant_id",
          "merchant_category_code",
          "amount",
          "currency",
          "channel",
          "is_international",
          "fraud_score",
          "risk_tier",
          "decision",
          "model_version_composite",
          "processing_ms",
          "pii_tokenised"
        ],
        "title": "TransactionResponse",
        "description": "Response schema for Transaction \u2014 safe to return (no raw PII)."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}